+++ src/components/BillList/PageNationSession.js
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from "react";
 import axios from "axios";
-import {Nav, Button} from "../style/StyledPageNationSession";
+import {Nav, Button} from "../../style/StyledPageNationSession";
 
 const KEY = process.env.REACT_APP_API_KEY;
 const BaseURL = `https://open.assembly.go.kr/portal/openapi/nzmimeepazxkubdpn?KEY=${KEY}&`;
